requirements.txt
fastapi
uvicorn
pygame
requests
paho-mqtt